split off dynmap json
